make the enums NSIntegers added bulk delete in the test

+++ Shared/Models/Generation.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-enum Generation: Int, Codable, CaseIterable {
+enum Generation: NSInteger, Codable, CaseIterable {
     case first = 1, second, third, fourth, fifth
 }
